fix: movie 리코일 값 업데이트 문제로 인한 delete 버튼 안 뜨는 에러
